docs(badge): add playground example
