+++ skills/carbon-react/components/badge.md
@@ -59,170 +59,3 @@ description: Carbon Badge component props and usage examples.
 }
 ```
 
-
-### With Children
-
-**Render**
-
-```tsx
-({ ...args }) => {
-  return (
-    <Badge id="badge-button" counter={99} {...args}>
-      <Button buttonType="secondary" aria-describedby="badge-button">
-        Filter
-      </Button>
-    </Badge>
-  );
-}
-```
-
-
-### Sizes
-
-**Render**
-
-```tsx
-({ ...args }) => {
-  return (
-    <>
-      <Badge id="badge-small" counter={99} size="small" {...args}>
-        <Icon type="alert" color="black" />
-      </Badge>
-      <Badge id="badge-medium" counter={99} size="medium" {...args} />
-      <Badge id="badge-large" counter={99} size="large" {...args} />
-    </>
-  );
-}
-```
-
-
-### Subtle Variant
-
-**Args**
-
-```tsx
-{
-  variant: "subtle",
-}
-```
-
-**Render**
-
-```tsx
-({ ...args }) => {
-  return (
-    <>
-      <Badge id="badge-subtle-small" counter={99} size="small" {...args}>
-        <Icon type="alert" color="black" />
-      </Badge>
-      <Badge id="badge-subtle-medium" counter={99} size="medium" {...args} />
-      <Badge id="badge-subtle-large" counter={99} size="large" {...args} />
-    </>
-  );
-}
-```
-
-
-### Inverse
-
-**Args**
-
-```tsx
-{
-  inverse: true,
-}
-```
-
-**Render**
-
-```tsx
-({ ...args }) => {
-  return (
-    <>
-      <Badge id="badge-inverse-small" counter={99} size="small" {...args}>
-        <Icon type="alert" color="white" />
-      </Badge>
-      <Badge id="badge-inverse-medium" counter={99} size="medium" {...args} />
-      <Badge id="badge-inverse-large" counter={99} size="large" {...args} />
-
-      <Badge
-        id="badge-icon"
-        counter={99}
-        size="small"
-        variant="subtle"
-        {...args}
-      >
-        <Icon type="alert" color="white" />
-      </Badge>
-      <Badge
-        id="badge-subtle-inverse-medium"
-        counter={99}
-        size="medium"
-        variant="subtle"
-        {...args}
-      />
-      <Badge
-        id="badge-subtle-inverse-large"
-        counter={99}
-        size="large"
-        variant="subtle"
-        {...args}
-      />
-    </>
-  );
-}
-```
-
-
-### With OnClick
-
-**Render**
-
-```tsx
-({ ...args }) => {
-  const counter = 9;
-  return (
-    <Badge
-      id="badge-onclick"
-      counter={counter}
-      onClick={() => {}}
-      aria-label={`Remove ${counter} filters.`}
-      {...args}
-    >
-      <Button aria-describedby="badge-onclick" buttonType="secondary">
-        Filter
-      </Button>
-    </Badge>
-  );
-}
-```
-
-
-### Custom Color
-
-**Render**
-
-```tsx
-({ ...args }) => {
-  const counter = 9;
-  return (
-    <Badge
-      id="badge-custom-color"
-      counter={counter}
-      onClick={() => {}}
-      aria-label={`Remove ${counter} filters.`}
-      color="--colorsSemanticNegative500"
-      {...args}
-    >
-      <Button
-        aria-describedby="badge-custom-color"
-        buttonType="secondary"
-        destructive
-      >
-        Filter
-      </Button>
-    </Badge>
-  );
-}
-```
-
