+++ src/components/badge/badge.mdx
@@ -1,4 +1,4 @@
-import { ArgTypes, Meta, Canvas } from "@storybook/addon-docs/blocks";
+import { ArgTypes, Meta, Canvas, Controls } from "@storybook/addon-docs/blocks";
 
 import * as BadgeStories from "./badge.stories";
 
@@ -29,6 +29,22 @@ To use Badge component, import `Badge` and use as a standalone component or wrap
 import Badge from "carbon-react/lib/components/badge";
 ```
 
+## Designer Notes
+
+A badge is a small numerical indicator positioned next to interactive elements.
+It shows when something needs action or review. Typically used for notifications within navigation, and to show that filters have been applied.
+
+## Playground
+
+Use this interactive example to explore Badge props with Storybook controls.
+
+<Canvas of={BadgeStories.Playground} />
+
+<Controls
+  of={BadgeStories.Playground}
+  include={["counter", "size", "variant", "inverse", "withButton"]}
+/>
+
 ## Examples
 
 ### Default
@@ -41,44 +57,6 @@ If the value is a string, the displayed value will be limited to 4 characters, t
 If the `counter` prop is not provided or its value is `0`, the `Badge` will not render.
 
 <Canvas of={BadgeStories.Default} />
-
-### With Children
-
-To position a `Badge` relative to another component, for example a `Button`, you can pass it as a child of `Badge`.
-
-Please make sure you associate the `Badge` to the component it relates to, this can be done by setting the `Badge`'s `id` to the child component's `aria-describedby`.
-
-<Canvas of={BadgeStories.WithChildren} />
-
-### Sizes
-
-You can use the `size` prop to change the size of the `Badge` to "small", "medium" (default), or "large".
-
-The "small" size will not visually display the `counter` value, however the prop will still be needed for the badge to render.
-
-<Canvas of={BadgeStories.Sizes} />
-
-### Subtle Variant
-
-By default, the `Badge` is rendered as the "typical" variant, however you can use the `variant` prop to change the appearance to "subtle".
-
-<Canvas of={BadgeStories.SubtleVariant} />
-
-### Inverse
-
-You can use the `inverse` prop to render the `Badge` with the inverse color scheme.
-
-<Canvas of={BadgeStories.Inverse} />
-
-### (Deprecated) Badge with onClick
-
-<Canvas of={BadgeStories.WithOnClick} />
-
-### (Deprecated) Custom Color
-
-You can use the `color` prop to override the default color of the component.
-
-<Canvas of={BadgeStories.CustomColor} />
 
 ## Props
